+++ html/signup.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css" />
     <link rel="stylesheet" href="/src/globalStyles.css" />
     <link rel="stylesheet" href="/src/sign.css" />
-    <script defer src="/js/signup.js"></script>
+    <script defer type="module" src="/js/signup.js"></script>
   </head>
   <body>
     <header>
